chore: Move CLI options to respective subcommand modules
